CoreData-Zugriff über Hilfsklassen (Protocol) Code umstrukturiert. Unschöne Codestellen ausgebessert.

+++ BrainTrain/Games/ColoredWords/ColoredWordsGame.swift
@@ -13,7 +13,7 @@ class ColoredWordsGame: GameProtocol {
     private var viewController: ColoredWordsViewController
     let name: String = "ColoredWords"
     let description: String = "..."
-    var didEndGame: ((Int) -> ())?
+    var didEndGame: (() -> ())?
     var playTime = 10
